moved simple functions to the so_utils module
